pyproject.toml
[project]
name = "EVE_TEST_BOT"
version = "0.3.0"
description = "EVE_TEST_BOT"
readme = "README.md"
requires-python = ">=3.10, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.poetry]
name = "nonebot-plugin-eve-tool"
version = "0.3.0"
description = "A NoneBot based plugin provides information query services for EVE online."
authors = ["zifox666 <zifox666@gmail.com>"]
license = "GPL-3.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiomysql = "^0.2.0"
aiohttp = "^3.9.5"
jieba = "^0.42.1"
httpx = "^0.27.0"
tqdm = "^4.66.4"
qrcode = "^7.4.2"
matplotlib = "^3.9.0"
aiofiles = "^23.2.1"
nonebot-plugin-alconna = "^0.47.1"
nonebot-plugin-apscheduler = "^0.4.0"
nonebot-plugin-htmlrender = "^0.3.2"
nonebot-plugin-waiter = "^0.6.2"
nonebot_adapter_onebot = "^2.3.1"
nonebot2 = "^2.3.1"
pillow = "^10.3.0"
websockets = "^12.0"
yarl = "^1.9.4"
zipp = "^3.19.2"
redis = "^5.0.6"
urllib3 = "^2.2.3"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
